Move chain sampling design phase props migration into its own module
Extracts the migration logic (oldSamplingDesignPhasePropKeys, the
phase-2-join-entity walk-up helpers, and migrateSamplingDesignPhaseProps
itself) out of the already-large chain manager index.js into a
dedicated chainSamplingDesignMigration.js. The public API is
unchanged: chain/index.js re-exports migrateSamplingDesignPhaseProps
from the new module, so no caller (surveyDataMigrationSteps.js,
chainsImportJob.js, the analysis manager barrel, or the integration
test) needs to change.
